fix: Node do not require window to build
