+++ packages/integrations/src/reportingobserver.ts
@@ -89,7 +89,7 @@ export class ReportingObserver implements Integration {
 
     this._getCurrentHub = getCurrentHub;
 
-    const observer = new (getGlobalObject()).ReportingObserver(this.handler.bind(this), {
+    const observer = new (getGlobalObject<any>()).ReportingObserver(this.handler.bind(this), {
       buffered: true,
       types: this._options.types,
     });
